Struct min_max_heap::MinMaxHeap

pub struct MinMaxHeap<T>(_);

A double-ended priority queue.

Most operations are O(log n).

Methods

impl<T> MinMaxHeap<T>

pub fn new() -> Self

Creates a new, empty MinMaxHeap.

O(1).

pub fn with_capacity(len: usize) -> Self

Creates a new, empty MinMaxHeap with space allocated to hold len elements.

O(n).

pub fn len(&self) -> usize

The number of elements in the heap.

O(1).

pub fn is_empty(&self) -> bool

Is the heap empty?

O(1).

impl<T: Ord> MinMaxHeap<T>

pub fn push(&mut self, element: T)

Adds an element to the heap.

Amortized O(log n); worst-case O(n) when the backing vector needs to grow.

pub fn peek_min(&self) -> Option<&T>

Gets a reference to the minimum element, if any.

O(1).

pub fn peek_max(&self) -> Option<&T>

Gets a reference to the maximum element, if any.

O(1).

pub fn pop_min(&mut self) -> Option<T>

Removes the minimum element, if any.

O(log n).

pub fn pop_max(&mut self) -> Option<T>

Removes the maximum element, if any.

O(log n).

pub fn push_pop_min(&mut self, element: T) -> T

Pushes an element, then pops the minimum element.

Unlike a push followed by a pop, this combined operation will not allocate.

O(log n).

pub fn push_pop_max(&mut self, element: T) -> T

Pushes an element, then pops the maximum element in an optimized fashion.

Unlike a push followed by a pop, this combined operation will not allocate.

O(log n).

pub fn replace_min(&mut self, element: T) -> Option<T>

Pops the minimum, then pushes an element in an optimized fashion.

O(log n).

pub fn replace_max(&mut self, element: T) -> Option<T>

Pops the maximum, then pushes an element in an optimized fashion.

O(log n).

pub fn into_vec_asc(self) -> Vec<T>

Returns an ascending (sorted) vector, reusing the heap’s storage.

O(n log n).

pub fn into_vec_desc(self) -> Vec<T>

Returns an descending (sorted) vector, reusing the heap’s storage.

O(n log n).

impl<T> MinMaxHeap<T>

pub fn clear(&mut self)

Drops all items from the heap.

O(n)

pub fn capacity(&self) -> usize

The number of elements the heap can hold without reallocating.

O(1)

pub fn reserve_exact(&mut self, additional: usize)

Reserves the minimum capacity for exactly additional more elements to be inserted in the given MinMaxHeap.

O(n)

Panics

Panics if the new capacity overflows usize.

pub fn reserve(&mut self, additional: usize)

Reserves the minimum capacity for at least additional more elements to be inserted in the given MinMaxHeap.

O(n)

Panics

Panics if the new capacity overflows usize.

pub fn shrink_to_fit(&mut self)

Discards extra capacity.

O(n)

pub fn into_vec(self) -> Vec<T>

Consumes the MinMaxHeap and returns its elements in a vector in arbitrary order.

O(n)

Important traits for Iter<'a, T>
impl<'a, T> Iterator for Iter<'a, T> type Item = &'a T;

pub fn iter(&self) -> Iter<T>

Returns a borrowing iterator over the min-max-heap’s elements in arbitrary order.

O(1) on creation, and O(1) for each next() operation.

Important traits for Drain<'a, T>
impl<'a, T> Iterator for Drain<'a, T> type Item = T;

pub fn drain(&mut self) -> Drain<T>

Returns a draining iterator over the min-max-heap’s elements in arbitrary order.

O(1) on creation, and O(1) for each next() operation.

Important traits for DrainAsc<'a, T>
impl<'a, T: Ord> Iterator for DrainAsc<'a, T> type Item = T;

pub fn drain_asc(&mut self) -> DrainAsc<T>

Returns a draining iterator over the min-max-heap’s elements in ascending (min-first) order.

O(1) on creation, and O(log n) for each next() operation.

Important traits for DrainDesc<'a, T>
impl<'a, T: Ord> Iterator for DrainDesc<'a, T> type Item = T;

pub fn drain_desc(&mut self) -> DrainDesc<T>

Returns a draining iterator over the min-max-heap’s elements in descending (max-first) order.

O(1) on creation, and O(log n) for each next() operation.