Struct min_max_heap::MinMaxHeap

pub struct MinMaxHeap<T>(_);

A double-ended priority queue.

Methods

impl<T> MinMaxHeap<T>

pub fn new() -> Self

Creates a new, empty MinMaxHeap.

pub fn with_capacity(len: usize) -> Self

Creates a new, empty MinMaxHeap with space allocated to hold len elements.

pub fn len(&self) -> usize

The number of elements in the heap.

pub fn is_empty(&self) -> bool

Is the heap empty?

impl<T: Ord> MinMaxHeap<T>

pub fn push(&mut self, element: T)

Adds an element to the heap.

pub fn peek_min(&self) -> Option<&T>

Gets a reference to the minimum element, if any.

pub fn peek_max(&self) -> Option<&T>

Gets a reference to the maximum element, if any.

pub fn pop_min(&mut self) -> Option<T>

Removes the minimum element, if any.

pub fn pop_max(&mut self) -> Option<T>

Removes the maximum element, if any.

pub fn push_pop_min(&mut self, element: T) -> T

Pushes an element, then pops the minimum element.

Unlike a push followed by a pop, this combined operation will not allocate.

pub fn push_pop_max(&mut self, element: T) -> T

Pushes an element, then pops the maximum element in an optimized fashion.

Unlike a push followed by a pop, this combined operation will not allocate.

pub fn replace_min(&mut self, element: T) -> Option<T>

Pops the minimum, then pushes an element in an optimized fashion.

pub fn replace_max(&mut self, element: T) -> Option<T>

Pops the maximum, then pushes an element in an optimized fashion.

pub fn into_vec_asc(self) -> Vec<T>

Returns an ascending (sorted) vector, reusing the heap’s storage.

pub fn into_vec_desc(self) -> Vec<T>

Returns an descending (sorted) vector, reusing the heap’s storage.

impl<T> MinMaxHeap<T>

pub fn clear(&mut self)

Drops all items from the heap.

pub fn capacity(&self) -> usize

The number of elements the heap can hold without reallocating.

pub fn reserve_exact(&mut self, additional: usize)

Reserves the minimum capacity for exactly additional more elements to be inserted in the given MinMaxHeap.

Panics

Panics if the new capacity overflows usize.

pub fn reserve(&mut self, additional: usize)

Reserves the minimum capacity for at least additional more elements to be inserted in the given MinMaxHeap.

Panics

Panics if the new capacity overflows usize.

pub fn shrink_to_fit(&mut self)

Discards extra capacity.

pub fn into_vec(self) -> Vec<T>

Consumes the MinMaxHeap and returns its elements in a vector in arbitrary order.

Important traits for Iter<'a, T>
impl<'a, T> Iterator for Iter<'a, T> type Item = &'a T;

pub fn iter(&self) -> Iter<T>

Returns a borrowing iterator over the min-max-heap’s elements in arbitrary order.

Important traits for Drain<'a, T>
impl<'a, T> Iterator for Drain<'a, T> type Item = T;

pub fn drain(&mut self) -> Drain<T>

Returns a draining iterator over the min-max-heap’s elements in arbitrary order.